+++ cmd/remarquee/cmds/cloud/ls.go
@@ -15,6 +15,7 @@ import (
 	"github.com/go-go-golems/glazed/pkg/middlewares"
 	"github.com/go-go-golems/glazed/pkg/settings"
 	"github.com/go-go-golems/glazed/pkg/types"
+	"github.com/go-go-golems/remarquee/cmd/remarquee/internal/appconfig"
 	"github.com/juruen/rmapi/model"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
@@ -228,10 +229,7 @@ func NewLsCobraCommand() (*cobra.Command, error) {
 	return cli.BuildCobraCommand(cmd,
 		cli.WithDualMode(true),
 		cli.WithGlazeToggleFlag("with-glaze-output"),
-		cli.WithParserConfig(cli.CobraParserConfig{
-			ShortHelpSections: []string{schema.DefaultSlug},
-			MiddlewaresFunc:   cli.CobraCommandDefaultMiddlewares,
-		}),
+		cli.WithParserConfig(appconfig.DefaultParserConfig()),
 	)
 }